change header too..

+++ src/mat_mult.h
@@ -1,2 +1,2 @@
 
-void gpu_blas_mmult(const float *A, const float *B, float *C, const int m, const int k, const int n);
+void gpu_blas_mmult(const double *A, const double *B, double *C, const int m, const int k, const int n);
